feat: add Ollama local LLM support alongside Gemini
Introduces a pluggable LLM provider system so users can run the app
fully offline using a local Ollama server instead of Google Gemini.

- Add src/services/gemini.service.js (extracted Gemini implementation)
- Add src/services/ollama.service.js (Ollama /api/chat client with vision
  support for screenshot analysis and conversation history)
- Convert src/services/llm.service.js into a dynamic provider router that
  delegates to gemini or ollama based on LLM_PROVIDER env var / runtime
  config — no restart required when switching
- Extend src/core/config.js with llm.provider and llm.ollama config block
- Add LLM_PROVIDER, OLLAMA_BASE_URL, OLLAMA_MODEL, OLLAMA_VISION_MODEL to
  env.example
- Add "AI Provider" selector and Ollama settings section to settings.html;
  Gemini/Ollama panels toggle based on selection
- Wire up new IPC handlers in main.js: get/set-llm-provider,
  get-ollama-status, test-ollama-connection
- Expose new IPC methods in preload.js
