+++ env.example
@@ -1,6 +1,20 @@
-# Google Gemini API Configuration
-# Get your API key from: https://makersuite.google.com/app/apikey
+# ── AI Provider ──────────────────────────────────────────────────────────────
+# Choose your LLM backend: 'gemini' (cloud, default) or 'ollama' (local)
+LLM_PROVIDER=gemini
+
+# ── Google Gemini (cloud) ─────────────────────────────────────────────────────
+# Required when LLM_PROVIDER=gemini
+# Get your API key from: https://aistudio.google.com/
 GEMINI_API_KEY=your_gemini_api_key_here
+
+# ── Ollama (local) ────────────────────────────────────────────────────────────
+# Required when LLM_PROVIDER=ollama
+# Install Ollama from https://ollama.ai, then:
+#   ollama pull llama3.2        # text model
+#   ollama pull llava           # vision model (for screenshot analysis)
+OLLAMA_BASE_URL=http://localhost:11434
+OLLAMA_MODEL=llama3.2
+OLLAMA_VISION_MODEL=llava
 
 # Speech Recognition Configuration
 # Choose one provider: azure or whisper
