+++ src/services/gemini.service.js
@@ -1,0 +1,973 @@
+const { GoogleGenerativeAI } = require('@google/generative-ai');
+const logger = require('../core/logger').createServiceLogger('LLM');
+const config = require('../core/config');
+const { promptLoader } = require('../../prompt-loader');
+
+class GeminiService {
+  constructor() {
+    this.client = null;
+    this.model = null;
+    this.isInitialized = false;
+    this.requestCount = 0;
+    this.errorCount = 0;
+
+    this.initializeClient();
+  }
+
+  initializeClient() {
+    const apiKey = config.getApiKey('GEMINI');
+
+    if (!apiKey || apiKey === 'your-api-key-here') {
+      logger.warn('Gemini API key not configured', {
+        keyExists: !!apiKey,
+        isPlaceholder: apiKey === 'your-api-key-here'
+      });
+      return;
+    }
+
+    try {
+      this.client = new GoogleGenerativeAI(apiKey);
+
+      const modelName = config.get('llm.gemini.model');
+      this.model = this.client.getGenerativeModel({
+        model: modelName,
+        generationConfig: this.getGenerationConfig()
+      });
+      this.isInitialized = true;
+
+      logger.info('Gemini AI client initialized successfully', {
+        model: modelName
+      });
+    } catch (error) {
+      logger.error('Failed to initialize Gemini client', {
+        error: error.message
+      });
+    }
+  }
+
+  getGenerationConfig(overrides = {}) {
+    const defaults = config.get('llm.gemini.generation') || {};
+    const fallback = {
+      temperature: 0.7,
+      topK: 40,
+      topP: 0.95,
+      maxOutputTokens: 4096
+    };
+
+    const merged = { ...fallback, ...defaults, ...overrides };
+    return Object.fromEntries(
+      Object.entries(merged).filter(([, value]) => value !== undefined && value !== null)
+    );
+  }
+
+  applyGenerationDefaults(request, overrides = {}) {
+    request.generationConfig = this.getGenerationConfig({ ...(request.generationConfig || {}), ...overrides });
+    return request;
+  }
+
+  extractTextFromCandidates(response) {
+    const candidates = Array.isArray(response?.candidates)
+      ? response.candidates
+      : Array.isArray(response)
+        ? response
+        : [];
+
+    if (!candidates.length) {
+      throw new Error('No candidates in Gemini response');
+    }
+
+    const candidateWithText = candidates.find(candidate => {
+      const parts = candidate?.content?.parts;
+      return Array.isArray(parts) && parts.some(part => typeof part.text === 'string' && part.text.trim().length > 0);
+    });
+
+    if (!candidateWithText) {
+      const finishReasons = candidates.map(c => c.finishReason || 'unknown').join(', ');
+      throw new Error(`No text parts in candidates. Finish reasons: ${finishReasons}`);
+    }
+
+    const textParts = candidateWithText.content.parts
+      .filter(part => typeof part.text === 'string' && part.text.trim().length > 0)
+      .map(part => part.text.trim());
+
+    if (!textParts.length) {
+      throw new Error(`Candidate parts missing text after filtering: ${JSON.stringify(candidateWithText)}`);
+    }
+
+    const text = textParts.join('\n');
+
+    return {
+      text,
+      candidate: candidateWithText,
+      finishReason: candidateWithText.finishReason || null
+    };
+  }
+
+  async processImageWithSkill(imageBuffer, mimeType, activeSkill, sessionMemory = [], programmingLanguage = null) {
+    if (!this.isInitialized) {
+      throw new Error('LLM service not initialized. Check Gemini API key configuration.');
+    }
+
+    if (!imageBuffer || !Buffer.isBuffer(imageBuffer)) {
+      throw new Error('Invalid image buffer provided to processImageWithSkill');
+    }
+
+    const startTime = Date.now();
+    this.requestCount++;
+
+    try {
+      const { promptLoader } = require('../../prompt-loader');
+      const skillPrompt = promptLoader.getSkillPrompt(activeSkill, programmingLanguage) || '';
+
+      const base64 = imageBuffer.toString('base64');
+
+      const request = {
+        contents: [
+          {
+            role: 'user',
+            parts: [
+              { text: this.formatImageInstruction(activeSkill, programmingLanguage) },
+              { inlineData: { data: base64, mimeType } }
+            ]
+          }
+        ]
+      };
+
+      this.applyGenerationDefaults(request);
+
+      if (skillPrompt && skillPrompt.trim().length > 0) {
+        request.systemInstruction = { parts: [{ text: skillPrompt }] };
+      }
+
+      let responseText;
+      const preferAlternative = !!config.get('llm.gemini.enableFallbackMethod');
+      try {
+        if (preferAlternative) {
+          logger.debug('Attempting alternative HTTPS method first for reliability');
+          responseText = await this.executeAlternativeRequest(request);
+        } else {
+          responseText = await this.executeRequest(request);
+        }
+      } catch (error) {
+        const secondaryLabel = preferAlternative ? 'primary SDK method' : 'alternative HTTPS method';
+        logger.warn(`${preferAlternative ? 'Alternative' : 'Primary'} method failed, trying ${secondaryLabel}`, { error: error.message });
+        const secondaryFn = preferAlternative ? this.executeRequest.bind(this) : this.executeAlternativeRequest.bind(this);
+
+        try {
+          responseText = await secondaryFn(request);
+        } catch (secondaryError) {
+          logger.error('Both Gemini request methods failed', {
+            firstError: error.message,
+            secondError: secondaryError.message
+          });
+          throw secondaryError;
+        }
+      }
+
+      const finalResponse = programmingLanguage
+        ? this.enforceProgrammingLanguage(responseText, programmingLanguage)
+        : responseText;
+
+      logger.logPerformance('LLM image processing', startTime, {
+        activeSkill,
+        imageSize: imageBuffer.length,
+        responseLength: finalResponse.length,
+        programmingLanguage: programmingLanguage || 'not specified',
+        requestId: this.requestCount
+      });
+
+      return {
+        response: finalResponse,
+        metadata: {
+          skill: activeSkill,
+          programmingLanguage,
+          processingTime: Date.now() - startTime,
+          requestId: this.requestCount,
+          usedFallback: false,
+          isImageAnalysis: true,
+          mimeType,
+          provider: 'gemini'
+        }
+      };
+    } catch (error) {
+      this.errorCount++;
+      logger.error('LLM image processing failed', {
+        error: error.message,
+        activeSkill,
+        requestId: this.requestCount
+      });
+
+      if (config.get('llm.gemini.fallbackEnabled')) {
+        return this.generateFallbackResponse('[image]', activeSkill);
+      }
+      throw error;
+    }
+  }
+
+  formatImageInstruction(activeSkill, programmingLanguage) {
+    const langNote = programmingLanguage ? ` Use only ${programmingLanguage.toUpperCase()} for any code.` : '';
+    return `Analyze this image for a ${activeSkill.toUpperCase()} question. Extract the problem concisely and provide the best possible solution with explanation and final code.${langNote}`;
+  }
+
+  async processTextWithSkill(text, activeSkill, sessionMemory = [], programmingLanguage = null) {
+    if (!this.isInitialized) {
+      throw new Error('LLM service not initialized. Check Gemini API key configuration.');
+    }
+
+    const startTime = Date.now();
+    this.requestCount++;
+
+    try {
+      logger.info('Processing text with LLM', {
+        activeSkill,
+        textLength: text.length,
+        hasSessionMemory: sessionMemory.length > 0,
+        programmingLanguage: programmingLanguage || 'not specified',
+        requestId: this.requestCount
+      });
+
+      const geminiRequest = this.buildGeminiRequest(text, activeSkill, sessionMemory, programmingLanguage);
+
+      const preferAlternative = !!config.get('llm.gemini.enableFallbackMethod');
+      let response;
+      try {
+        if (preferAlternative) {
+          logger.debug('Attempting alternative HTTPS method first for text processing');
+          response = await this.executeAlternativeRequest(geminiRequest);
+        } else {
+          response = await this.executeRequest(geminiRequest);
+        }
+      } catch (error) {
+        const secondaryLabel = preferAlternative ? 'primary SDK method' : 'alternative HTTPS method';
+        logger.warn(`${preferAlternative ? 'Alternative' : 'Primary'} method failed, trying ${secondaryLabel}`, {
+          error: error.message,
+          requestId: this.requestCount
+        });
+        const secondaryFn = preferAlternative ? this.executeRequest.bind(this) : this.executeAlternativeRequest.bind(this);
+        try {
+          response = await secondaryFn(geminiRequest);
+        } catch (secondaryError) {
+          logger.error('Both Gemini request methods failed for text processing', {
+            firstError: error.message,
+            secondError: secondaryError.message,
+            requestId: this.requestCount
+          });
+          throw secondaryError;
+        }
+      }
+
+      const finalResponse = programmingLanguage
+        ? this.enforceProgrammingLanguage(response, programmingLanguage)
+        : response;
+
+      logger.logPerformance('LLM text processing', startTime, {
+        activeSkill,
+        textLength: text.length,
+        responseLength: finalResponse.length,
+        programmingLanguage: programmingLanguage || 'not specified',
+        requestId: this.requestCount
+      });
+
+      return {
+        response: finalResponse,
+        metadata: {
+          skill: activeSkill,
+          programmingLanguage,
+          processingTime: Date.now() - startTime,
+          requestId: this.requestCount,
+          usedFallback: false,
+          provider: 'gemini'
+        }
+      };
+    } catch (error) {
+      this.errorCount++;
+      logger.error('LLM processing failed', {
+        error: error.message,
+        activeSkill,
+        programmingLanguage: programmingLanguage || 'not specified',
+        requestId: this.requestCount
+      });
+
+      if (config.get('llm.gemini.fallbackEnabled')) {
+        return this.generateFallbackResponse(text, activeSkill);
+      }
+
+      throw error;
+    }
+  }
+
+  async processTranscriptionWithIntelligentResponse(text, activeSkill, sessionMemory = [], programmingLanguage = null) {
+    if (!this.isInitialized) {
+      throw new Error('LLM service not initialized. Check Gemini API key configuration.');
+    }
+
+    const startTime = Date.now();
+    this.requestCount++;
+
+    try {
+      logger.info('Processing transcription with intelligent response', {
+        activeSkill,
+        textLength: text.length,
+        hasSessionMemory: sessionMemory.length > 0,
+        programmingLanguage: programmingLanguage || 'not specified',
+        requestId: this.requestCount
+      });
+
+      const geminiRequest = this.buildIntelligentTranscriptionRequest(text, activeSkill, sessionMemory, programmingLanguage);
+
+      const preferAlternative = !!config.get('llm.gemini.enableFallbackMethod');
+      let response;
+      try {
+        if (preferAlternative) {
+          logger.debug('Attempting alternative HTTPS method first for transcription processing');
+          response = await this.executeAlternativeRequest(geminiRequest);
+        } else {
+          response = await this.executeRequest(geminiRequest);
+        }
+      } catch (error) {
+        const secondaryLabel = preferAlternative ? 'primary SDK method' : 'alternative HTTPS method';
+        logger.warn(`${preferAlternative ? 'Alternative' : 'Primary'} method failed, trying ${secondaryLabel}`, {
+          error: error.message,
+          requestId: this.requestCount
+        });
+        const secondaryFn = preferAlternative ? this.executeRequest.bind(this) : this.executeAlternativeRequest.bind(this);
+        try {
+          response = await secondaryFn(geminiRequest);
+        } catch (secondaryError) {
+          logger.error('Both Gemini request methods failed for transcription processing', {
+            firstError: error.message,
+            secondError: secondaryError.message,
+            requestId: this.requestCount
+          });
+          throw secondaryError;
+        }
+      }
+
+      const finalResponse = programmingLanguage
+        ? this.enforceProgrammingLanguage(response, programmingLanguage)
+        : response;
+
+      logger.logPerformance('LLM transcription processing', startTime, {
+        activeSkill,
+        textLength: text.length,
+        responseLength: finalResponse.length,
+        programmingLanguage: programmingLanguage || 'not specified',
+        requestId: this.requestCount
+      });
+
+      return {
+        response: finalResponse,
+        metadata: {
+          skill: activeSkill,
+          programmingLanguage,
+          processingTime: Date.now() - startTime,
+          requestId: this.requestCount,
+          usedFallback: false,
+          isTranscriptionResponse: true,
+          provider: 'gemini'
+        }
+      };
+    } catch (error) {
+      this.errorCount++;
+      logger.error('LLM transcription processing failed', {
+        error: error.message,
+        activeSkill,
+        programmingLanguage: programmingLanguage || 'not specified',
+        requestId: this.requestCount
+      });
+
+      if (config.get('llm.gemini.fallbackEnabled')) {
+        return this.generateIntelligentFallbackResponse(text, activeSkill);
+      }
+
+      throw error;
+    }
+  }
+
+  enforceProgrammingLanguage(text, programmingLanguage) {
+    try {
+      if (!text || !programmingLanguage) return text;
+      const norm = String(programmingLanguage).toLowerCase();
+      const fenceTagMap = { cpp: 'cpp', c: 'c', python: 'python', java: 'java', javascript: 'javascript', js: 'javascript' };
+      const fenceTag = fenceTagMap[norm] || norm || 'text';
+
+      const replacedBackticks = text.replace(/```([^\n]*)\n/g, (match, info) => {
+        const current = (info || '').trim();
+        if (current.split(/\s+/)[0].toLowerCase() === fenceTag) return match;
+        return '```' + fenceTag + '\n';
+      });
+
+      const normalizedTildes = replacedBackticks.replace(/~~~([^\n]*)\n/g, () => '```' + fenceTag + '\n');
+
+      return normalizedTildes;
+    } catch (_) {
+      return text;
+    }
+  }
+
+  buildGeminiRequest(text, activeSkill, sessionMemory, programmingLanguage) {
+    const sessionManager = require('../managers/session.manager');
+
+    if (sessionManager && typeof sessionManager.getConversationHistory === 'function') {
+      const conversationHistory = sessionManager.getConversationHistory(15);
+      const skillContext = sessionManager.getSkillContext(activeSkill, programmingLanguage);
+      return this.buildGeminiRequestWithHistory(text, activeSkill, conversationHistory, skillContext, programmingLanguage);
+    }
+
+    const requestComponents = promptLoader.getRequestComponents(
+      activeSkill,
+      text,
+      sessionMemory,
+      programmingLanguage
+    );
+
+    const request = {
+      contents: []
+    };
+
+    this.applyGenerationDefaults(request);
+
+    if (requestComponents.shouldUseModelMemory && requestComponents.skillPrompt) {
+      request.systemInstruction = {
+        parts: [{ text: requestComponents.skillPrompt }]
+      };
+    }
+
+    request.contents.push({
+      role: 'user',
+      parts: [{ text: this.formatUserMessage(text, activeSkill) }]
+    });
+
+    return request;
+  }
+
+  buildGeminiRequestWithHistory(text, activeSkill, conversationHistory, skillContext, programmingLanguage) {
+    const request = {
+      contents: []
+    };
+
+    this.applyGenerationDefaults(request);
+
+    if (skillContext.skillPrompt) {
+      request.systemInstruction = {
+        parts: [{ text: skillContext.skillPrompt }]
+      };
+    }
+
+    const conversationContents = conversationHistory
+      .filter(event => {
+        return event.role !== 'system' &&
+               event.content &&
+               typeof event.content === 'string' &&
+               event.content.trim().length > 0;
+      })
+      .map(event => {
+        const content = event.content.trim();
+        return {
+          role: event.role === 'model' ? 'model' : 'user',
+          parts: [{ text: content }]
+        };
+      });
+
+    request.contents.push(...conversationContents);
+
+    const formattedMessage = this.formatUserMessage(text, activeSkill);
+    if (!formattedMessage || formattedMessage.trim().length === 0) {
+      throw new Error('Failed to format user message or message is empty');
+    }
+
+    request.contents.push({
+      role: 'user',
+      parts: [{ text: formattedMessage }]
+    });
+
+    return request;
+  }
+
+  buildIntelligentTranscriptionRequest(text, activeSkill, sessionMemory, programmingLanguage) {
+    const cleanText = text && typeof text === 'string' ? text.trim() : '';
+    if (!cleanText) {
+      throw new Error('Empty or invalid transcription text provided to buildIntelligentTranscriptionRequest');
+    }
+
+    const sessionManager = require('../managers/session.manager');
+
+    if (sessionManager && typeof sessionManager.getConversationHistory === 'function') {
+      const conversationHistory = sessionManager.getConversationHistory(10);
+      const skillContext = sessionManager.getSkillContext(activeSkill, programmingLanguage);
+      return this.buildIntelligentTranscriptionRequestWithHistory(cleanText, activeSkill, conversationHistory, skillContext, programmingLanguage);
+    }
+
+    const request = {
+      contents: []
+    };
+
+    this.applyGenerationDefaults(request);
+
+    const intelligentPrompt = this.getIntelligentTranscriptionPrompt(activeSkill, programmingLanguage);
+    if (!intelligentPrompt) {
+      throw new Error('Failed to generate intelligent transcription prompt');
+    }
+
+    request.systemInstruction = {
+      parts: [{ text: intelligentPrompt }]
+    };
+
+    request.contents.push({
+      role: 'user',
+      parts: [{ text: cleanText }]
+    });
+
+    return request;
+  }
+
+  buildIntelligentTranscriptionRequestWithHistory(text, activeSkill, conversationHistory, skillContext, programmingLanguage) {
+    const request = {
+      contents: []
+    };
+
+    this.applyGenerationDefaults(request);
+
+    const intelligentPrompt = this.getIntelligentTranscriptionPrompt(activeSkill, programmingLanguage);
+    request.systemInstruction = { parts: [{ text: intelligentPrompt }] };
+
+    const conversationContents = conversationHistory
+      .filter(event => {
+        return event.role !== 'system' &&
+               event.content &&
+               typeof event.content === 'string' &&
+               event.content.trim().length > 0;
+      })
+      .slice(-8)
+      .map(event => {
+        const content = event.content.trim();
+        if (!content) return null;
+        return {
+          role: event.role === 'model' ? 'model' : 'user',
+          parts: [{ text: content }]
+        };
+      })
+      .filter(content => content !== null);
+
+    request.contents.push(...conversationContents);
+
+    const cleanText = text && typeof text === 'string' ? text.trim() : '';
+    if (!cleanText) {
+      throw new Error('Empty or invalid transcription text provided');
+    }
+
+    request.contents.push({
+      role: 'user',
+      parts: [{ text: cleanText }]
+    });
+
+    if (request.contents.length === 0) {
+      throw new Error('No valid content to send to Gemini API');
+    }
+
+    return request;
+  }
+
+  getIntelligentTranscriptionPrompt(activeSkill, programmingLanguage) {
+    let prompt = `# Intelligent Transcription Response System
+
+Assume you are asked a question in ${activeSkill.toUpperCase()} mode. Your job is to intelligently respond to question/message with appropriate brevity.
+Assume you are in an interview and you need to perform best in ${activeSkill.toUpperCase()} mode.
+Always respond to the point, do not repeat the question or unnecessary information which is not related to ${activeSkill}.`;
+
+    if (programmingLanguage) {
+      const lang = String(programmingLanguage).toLowerCase();
+      const languageMap = { cpp: 'C++', c: 'C', python: 'Python', java: 'Java', javascript: 'JavaScript', js: 'JavaScript' };
+      const fenceTagMap = { cpp: 'cpp', c: 'c', python: 'python', java: 'java', javascript: 'javascript', js: 'javascript' };
+      const languageTitle = languageMap[lang] || (lang.charAt(0).toUpperCase() + lang.slice(1));
+      const fenceTag = fenceTagMap[lang] || lang || 'text';
+      prompt += `\n\nCODING CONTEXT: Respond ONLY in ${languageTitle}. All code blocks must use triple backticks with language tag \`\`\`${fenceTag}\`\`\`. Do not include other languages unless explicitly asked.`;
+    }
+
+    prompt += `
+
+## Response Rules:
+
+### If the transcription is casual conversation, greetings, or NOT related to ${activeSkill}:
+- Respond with: "Yeah, I'm listening. Ask your question relevant to ${activeSkill}."
+- Or similar brief acknowledgments like: "I'm here, what's your ${activeSkill} question?"
+
+### If the transcription IS relevant to ${activeSkill} or is a follow-up question:
+- Provide a comprehensive, detailed response
+- Use bullet points, examples, and explanations
+- Focus on actionable insights and complete answers
+- Do not truncate or shorten your response
+
+### Examples of casual/irrelevant messages:
+- "Hello", "Hi there", "How are you?"
+- "What's the weather like?"
+- "I'm just testing this"
+- Random conversations not related to ${activeSkill}
+
+### Examples of relevant messages:
+- Actual questions about ${activeSkill} concepts
+- Follow-up questions to previous responses
+- Requests for clarification on ${activeSkill} topics
+- Problem-solving requests related to ${activeSkill}
+
+## Response Format:
+- Keep responses detailed
+- Use bullet points for structured answers
+- Be encouraging and helpful
+- Stay focused on ${activeSkill}
+
+If the user's input is a coding or DSA problem statement and contains no code, produce a complete, runnable solution in the selected programming language without asking for more details. Always include the final implementation in a properly tagged code block.
+
+Remember: Be intelligent about filtering - only provide detailed responses when the user actually needs help with ${activeSkill}.`;
+
+    return prompt;
+  }
+
+  formatUserMessage(text, activeSkill) {
+    return `Context: ${activeSkill.toUpperCase()} analysis request\n\nText to analyze:\n${text}`;
+  }
+
+  async executeRequest(geminiRequest) {
+    const maxRetries = config.get('llm.gemini.maxRetries');
+    const timeout = config.get('llm.gemini.timeout');
+
+    for (let attempt = 1; attempt <= maxRetries; attempt++) {
+      try {
+        await this.performPreflightCheck();
+
+        const timeoutPromise = new Promise((_, reject) =>
+          setTimeout(() => reject(new Error('Request timeout')), timeout)
+        );
+
+        const requestPromise = this.model.generateContent(geminiRequest);
+        const result = await Promise.race([requestPromise, timeoutPromise]);
+
+        if (!result.response) {
+          throw new Error('Empty response from Gemini API');
+        }
+
+        const { text, finishReason } = this.extractTextFromCandidates(result.response);
+
+        if (finishReason === 'MAX_TOKENS') {
+          logger.warn('Gemini primary response reached max tokens limit', { attempt, finishReason });
+        }
+
+        return text;
+      } catch (error) {
+        const errorInfo = this.analyzeError(error);
+
+        logger.warn(`Gemini API attempt ${attempt} failed`, {
+          error: error.message,
+          errorType: errorInfo.type,
+          remainingAttempts: maxRetries - attempt
+        });
+
+        if (attempt === maxRetries) {
+          const finalError = new Error(`Gemini API failed after ${maxRetries} attempts: ${error.message}`);
+          finalError.errorAnalysis = errorInfo;
+          finalError.originalError = error;
+          throw finalError;
+        }
+
+        const baseDelay = errorInfo.isNetworkError ? 2500 : 1500;
+        const delay = baseDelay * attempt + Math.random() * 1000;
+        await this.delay(delay);
+      }
+    }
+  }
+
+  async performPreflightCheck() {
+    try {
+      await this.testNetworkConnection({
+        host: 'generativelanguage.googleapis.com',
+        port: 443,
+        name: 'Gemini API Endpoint'
+      });
+    } catch (error) {
+      logger.warn('Preflight check failed', { error: error.message });
+    }
+  }
+
+  getUserAgent() {
+    try {
+      if (typeof navigator !== 'undefined' && navigator.userAgent) {
+        return navigator.userAgent;
+      }
+      return `Node.js/${process.version} (${process.platform}; ${process.arch})`;
+    } catch {
+      return 'Unknown';
+    }
+  }
+
+  analyzeError(error) {
+    const errorMessage = error.message.toLowerCase();
+
+    if (errorMessage.includes('fetch failed') ||
+        errorMessage.includes('network error') ||
+        errorMessage.includes('enotfound') ||
+        errorMessage.includes('econnrefused') ||
+        errorMessage.includes('timeout')) {
+      return { type: 'NETWORK_ERROR', isNetworkError: true, suggestedAction: 'Check internet connection and firewall settings' };
+    }
+
+    if (errorMessage.includes('unauthorized') ||
+        errorMessage.includes('invalid api key') ||
+        errorMessage.includes('forbidden')) {
+      return { type: 'AUTH_ERROR', isNetworkError: false, suggestedAction: 'Verify Gemini API key configuration' };
+    }
+
+    if (errorMessage.includes('quota') ||
+        errorMessage.includes('rate limit') ||
+        errorMessage.includes('too many requests')) {
+      return { type: 'RATE_LIMIT_ERROR', isNetworkError: false, suggestedAction: 'Wait before retrying or check API quota' };
+    }
+
+    if (errorMessage.includes('request timeout') || errorMessage.includes('etimedout')) {
+      return { type: 'TIMEOUT_ERROR', isNetworkError: true, suggestedAction: 'Check network latency or increase timeout' };
+    }
+
+    return { type: 'UNKNOWN_ERROR', isNetworkError: false, suggestedAction: 'Check logs for more details' };
+  }
+
+  async checkNetworkConnectivity() {
+    const connectivityTests = [
+      { host: 'google.com', port: 443, name: 'Google (HTTPS)' },
+      { host: 'generativelanguage.googleapis.com', port: 443, name: 'Gemini API Endpoint' }
+    ];
+
+    const results = await Promise.allSettled(
+      connectivityTests.map(test => this.testNetworkConnection(test))
+    );
+
+    const connectivity = {
+      timestamp: new Date().toISOString(),
+      tests: results.map((result, index) => ({
+        ...connectivityTests[index],
+        success: result.status === 'fulfilled' && result.value,
+        error: result.status === 'rejected' ? result.reason.message : null
+      }))
+    };
+
+    logger.info('Network connectivity check completed', connectivity);
+    return connectivity;
+  }
+
+  async testNetworkConnection({ host, port, name }) {
+    return new Promise((resolve, reject) => {
+      const net = require('net');
+      const socket = new net.Socket();
+
+      const timeout = setTimeout(() => {
+        socket.destroy();
+        reject(new Error(`Connection timeout to ${host}:${port}`));
+      }, 5000);
+
+      socket.on('connect', () => {
+        clearTimeout(timeout);
+        socket.destroy();
+        resolve(true);
+      });
+
+      socket.on('error', (error) => {
+        clearTimeout(timeout);
+        reject(new Error(`Connection failed to ${host}:${port}: ${error.message}`));
+      });
+
+      socket.connect(port, host);
+    });
+  }
+
+  generateFallbackResponse(text, activeSkill) {
+    logger.info('Generating fallback response', { activeSkill });
+
+    const fallbackResponses = {
+      'dsa': 'This appears to be a data structures and algorithms problem. Consider breaking it down into smaller components and identifying the appropriate algorithm or data structure to use.',
+      'system-design': 'For this system design question, consider scalability, reliability, and the trade-offs between different architectural approaches.',
+      'programming': 'This looks like a programming challenge. Focus on understanding the requirements, edge cases, and optimal time/space complexity.',
+      'default': 'I can help analyze this content. Please ensure your Gemini API key is properly configured for detailed analysis.'
+    };
+
+    const response = fallbackResponses[activeSkill] || fallbackResponses.default;
+
+    return {
+      response,
+      metadata: {
+        skill: activeSkill,
+        processingTime: 0,
+        requestId: this.requestCount,
+        usedFallback: true,
+        provider: 'gemini'
+      }
+    };
+  }
+
+  generateIntelligentFallbackResponse(text, activeSkill) {
+    logger.info('Generating intelligent fallback response for transcription', { activeSkill });
+
+    const skillKeywords = {
+      'dsa': ['algorithm', 'data structure', 'array', 'tree', 'graph', 'sort', 'search', 'complexity', 'big o'],
+      'programming': ['code', 'function', 'variable', 'class', 'method', 'bug', 'debug', 'syntax'],
+      'system-design': ['scalability', 'database', 'architecture', 'microservice', 'load balancer', 'cache'],
+      'behavioral': ['interview', 'experience', 'situation', 'leadership', 'conflict', 'team'],
+      'sales': ['customer', 'deal', 'negotiation', 'price', 'revenue', 'prospect'],
+      'presentation': ['slide', 'audience', 'public speaking', 'presentation', 'nervous'],
+      'data-science': ['data', 'model', 'machine learning', 'statistics', 'analytics', 'python', 'pandas'],
+      'devops': ['deployment', 'ci/cd', 'docker', 'kubernetes', 'infrastructure', 'monitoring'],
+      'negotiation': ['negotiate', 'compromise', 'agreement', 'terms', 'conflict resolution']
+    };
+
+    const textLower = text.toLowerCase();
+    const relevantKeywords = skillKeywords[activeSkill] || [];
+    const hasRelevantKeywords = relevantKeywords.some(keyword => textLower.includes(keyword));
+
+    const questionIndicators = ['how', 'what', 'why', 'when', 'where', 'can you', 'could you', 'should i', '?'];
+    const seemsLikeQuestion = questionIndicators.some(indicator => textLower.includes(indicator));
+
+    let response;
+    if (hasRelevantKeywords || seemsLikeQuestion) {
+      response = `I'm having trouble processing that right now, but it sounds like a ${activeSkill} question. Could you rephrase or ask more specifically about what you need help with?`;
+    } else {
+      response = `Yeah, I'm listening. Ask your question relevant to ${activeSkill}.`;
+    }
+
+    return {
+      response,
+      metadata: {
+        skill: activeSkill,
+        processingTime: 0,
+        requestId: this.requestCount,
+        usedFallback: true,
+        isTranscriptionResponse: true,
+        provider: 'gemini'
+      }
+    };
+  }
+
+  async testConnection() {
+    if (!this.isInitialized) {
+      return { success: false, error: 'Service not initialized', provider: 'gemini' };
+    }
+
+    try {
+      const networkCheck = await this.checkNetworkConnectivity();
+      const hasNetworkIssues = networkCheck.tests.some(test => !test.success);
+
+      if (hasNetworkIssues) {
+        logger.warn('Network connectivity issues detected', networkCheck);
+      }
+
+      const testRequest = {
+        contents: [{
+          role: 'user',
+          parts: [{ text: 'Test connection. Please respond with "OK".' }]
+        }]
+      };
+
+      this.applyGenerationDefaults(testRequest, { temperature: 0, maxOutputTokens: 10 });
+
+      const startTime = Date.now();
+      const result = await this.model.generateContent(testRequest);
+      const latency = Date.now() - startTime;
+      const { text } = this.extractTextFromCandidates(result.response);
+
+      return {
+        success: true,
+        response: text,
+        latency,
+        networkConnectivity: networkCheck,
+        provider: 'gemini'
+      };
+    } catch (error) {
+      const errorAnalysis = this.analyzeError(error);
+      return {
+        success: false,
+        error: error.message,
+        errorAnalysis,
+        networkConnectivity: await this.checkNetworkConnectivity().catch(() => null),
+        provider: 'gemini'
+      };
+    }
+  }
+
+  updateApiKey(newApiKey) {
+    process.env.GEMINI_API_KEY = newApiKey;
+    this.isInitialized = false;
+    this.initializeClient();
+    logger.info('API key updated and client reinitialized');
+  }
+
+  getStats() {
+    return {
+      isInitialized: this.isInitialized,
+      requestCount: this.requestCount,
+      errorCount: this.errorCount,
+      successRate: this.requestCount > 0 ? ((this.requestCount - this.errorCount) / this.requestCount) * 100 : 0,
+      config: config.get('llm.gemini'),
+      provider: 'gemini'
+    };
+  }
+
+  delay(ms) {
+    return new Promise(resolve => setTimeout(resolve, ms));
+  }
+
+  async executeAlternativeRequest(geminiRequest) {
+    const https = require('https');
+    const apiKey = config.getApiKey('GEMINI');
+    const model = config.get('llm.gemini.model');
+
+    logger.info('Using alternative HTTPS request method');
+
+    const url = `https://generativelanguage.googleapis.com/v1beta/models/${model}:generateContent`;
+    const postData = JSON.stringify(geminiRequest);
+    const agent = new https.Agent({ keepAlive: true, maxSockets: 1 });
+
+    const options = {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+        'x-goog-api-key': apiKey,
+        'Content-Length': Buffer.byteLength(postData),
+        'User-Agent': this.getUserAgent()
+      },
+      timeout: config.get('llm.gemini.timeout'),
+      agent
+    };
+
+    return new Promise((resolve, reject) => {
+      const req = https.request(url, options, (res) => {
+        let data = '';
+
+        res.on('data', (chunk) => { data += chunk; });
+
+        res.on('end', () => {
+          try {
+            if (res.statusCode !== 200) {
+              reject(new Error(`HTTP ${res.statusCode}: ${data}`));
+              return;
+            }
+
+            const response = JSON.parse(data);
+            const { text, finishReason } = this.extractTextFromCandidates(response);
+
+            if (finishReason === 'MAX_TOKENS') {
+              logger.warn('Gemini alternative response reached max tokens limit', { finishReason });
+            }
+
+            resolve(text.trim());
+          } catch (parseError) {
+            reject(new Error(`Failed to parse response: ${parseError.message}`));
+          }
+        });
+      });
+
+      req.on('error', (error) => { reject(new Error(`Alternative request failed: ${error.message}`)); });
+      req.on('timeout', () => { req.destroy(); reject(new Error('Alternative request timeout')); });
+
+      req.write(postData);
+      req.end();
+    });
+  }
+}
+
+module.exports = new GeminiService();
